TA-SQL/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==40.0.2
cycler==0.12.1
distro==1.9.0
docutils==0.21.2
fonttools==4.58.4
frozenlist==1.7.0
func_timeout==4.3.5
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lark==1.2.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mo-dots==9.189.22201
mo-future==6.2.21303
mo-imports==7.187.22201
mo-parsing==8.204.22252
mo-sql-parsing==8.205.22260
msal==1.20.0
multidict==6.5.0
mysql-connector-python==9.3.0
numpy==1.26.4
openai==1.93.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
roman-numerals-py==3.1.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tqdm==4.65.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
-e git+https://github.com/rklop/SURF2025.git@bf44a196d3b0d1ac6e22e4533a58470953971754#egg=verieql
wcwidth==0.2.13
wheel==0.45.1
yarl==1.20.1
z3-solver==4.15.1.0
